+++ lib/Steps/PromptForParameters.ts
@@ -47,6 +47,7 @@ export class PromptForParameters extends BaseStep {
     ]);
 
     url = this._getFullUrl(answers, organization.slug, project.slug);
+    const dsnKeyUrl = this._getDSNKeyUrl(answers, project.slug);
     const dsn = await prompt([
       {
         message: 'DSN:',
@@ -56,7 +57,7 @@ export class PromptForParameters extends BaseStep {
         validate: this._validateDSN,
         when: this._shouldAsk(answers, 'config.dsn.public', () => {
           dim('Please copy/paste your DSN');
-          dim(`It can be found here: ${url}`);
+          dim(`It can be found here: ${dsnKeyUrl}`);
         }),
       },
     ]);
@@ -105,6 +106,16 @@ export class PromptForParameters extends BaseStep {
       projectSlug || 'project_slug',
     );
     return `${baseUrl}${orgSlug}/${projSlug}`;
+  }
+
+  private _getDSNKeyUrl(answers: Answers, projectSlug?: string): string {
+    const baseUrl = this._argv.url;
+    const projSlug = _.get(
+      answers,
+      'config.project.slug',
+      projectSlug || 'project_slug',
+    );
+    return `${baseUrl}settings/projects/${projSlug}/keys`;
   }
 
   private _shouldAsk(